refactor: move scan/cache/clone/media under api/services (no behavior change)
Creates api/services/ package and relocates the four data-layer modules;
updates all intra-service relative imports, test import paths, and
mock.patch string targets to the new api.services.* namespace.
