+++ api/services/scan.py
@@ -25,7 +25,7 @@ from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Callable, Iterator
 
-from .env import env_bool
+from api.env import env_bool
 from .cache import (
     cache_load_files,
     cache_load_git_history,
@@ -33,7 +33,7 @@ from .cache import (
     cache_save_git_history,
 )
 from .media import probe_media_dims
-from .types import (
+from api.types import (
     BusynessThresholds,
     CommitEntry,
     DirNode,
