+++ api/tests/test_media.py
@@ -9,7 +9,7 @@ import zlib
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-from api.media import _parse_svg_length, probe_media_dims
+from api.services.media import _parse_svg_length, probe_media_dims
 
 
 def _write_minimal_png(path: Path, width: int, height: int) -> None:
